fix minor bug
millis() returns an unsigned long size which will overflow timer, a uint16, pretty quickly. changed the code to use the proper sized variable.

+++ examples/parsing/parsing.pde
@@ -94,7 +94,7 @@ void useInterrupt(boolean v) {
   }
 }
 
-uint16_t timer = millis();
+unsigned long timer = millis();
 void loop()                     // run over and over again
 {
   // in case you are not using the interrupt above, you'll
